Run cli benchmarks only using config files
